Tweak DecarcerateDane.com custom components
