+++ views/measure-sidebar.js
@@ -165,8 +165,6 @@ const measureVoteCounts = ({ measure, offices }) => {
     'CA': { Upper: 'Senate', Lower: 'Assembly' },
   }
 
-  const wiNotDane = measure.legislature_name === 'Dane County' ? 'WI' : ''
-
   return html`
     <div class="panel-block">
       <div style="width: 100%;">
@@ -197,7 +195,7 @@ const measureVoteCounts = ({ measure, offices }) => {
               <td class="has-text-right">Nay</td>
             </tr>
             <tr>
-              <td class="has-text-left has-text-grey">${wiNotDane || legislature_name.replace(' Congress', '')}</td>
+              <td class="has-text-left has-text-grey">${legislature_name.replace(' Congress', '')}</td>
               <td class="has-text-right">${yeas || 0}</td>
               <td class="has-text-right">${nays || 0}</td>
             </tr>
